% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_database.R
\name{calc_shrub_density}
\alias{calc_shrub_density}
\title{Calculate density of shrub}
\usage{
calc_shrub_density(width = 2, length = 4, height = 1)
}
\arguments{
\item{width}{Width of shrub (numeric, default 2)}

\item{length}{length of shrub (numeric, default 4)}

\item{height}{Height of shrub (numeric, default 1)}

\item{volume}{volume of shrub (numeric)}

\item{mass}{mass of shrub (numeric)}
}
\value{
density of shrub (numeric)
}
\description{
Calculate density of shrub
}
