% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\name{gpa}
\alias{gpa}
\title{Performs and visualizes Generalized Procrustes Analysis for shape data}
\usage{
gpa(sirens_dorsal)
}
\arguments{
\item{sirens_dorsal}{siren specimen shape dataset}
}
\value{
a vector containing the Centroid Size of the configurations (graphic)
}
\description{
Performs and visualizes Generalized Procrustes Analysis for shape data
}
