% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cva.R, R/pca.R
\name{pca}
\alias{pca}
\title{Performs and visualizes a Principal Components Analysis for shape data}
\usage{
pca(gpa)

pca(gpa)
}
\arguments{
\item{gpa}{General Procrustes Analysis for shape data}

\item{pca}{Principal Components Analysis for shape data}
}
\value{
PCA graph for shape data (graph)

PCA graph for shape data (graph)
}
\description{
Performs and visualizes a Principal Components Analysis for shape data

Performs and visualizes a Principal Components Analysis for shape data
}
