% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{sirens_map}
\alias{sirens_map}
\title{Generates map of specimen locality information}
\usage{
sirens_map(sirens, Longitude, Latitude)
}
\arguments{
\item{sirens}{siren specimen dataset}

\item{Longitude}{longitudinal coordinate (numeric)}

\item{Latitude}{latitudinal coordinate (numeric)}

\item{world}{the specific map used for referencing (graphic)}
}
\value{
map of specimen localities (graphic)
}
\description{
Generates map of specimen locality information
}
